paytracker/pyproject.toml
[project]
name = "paytracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyautogui>=0.9.54",
    "python-dotenv>=1.1.0",
]
